% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{add_prompt}
\alias{add_prompt}
\title{Add prompt to a data frame of preprints}
\usage{
add_prompt(preprints, ...)
}
\arguments{
\item{preprints}{Result from \code{\link[=get_preprints]{get_preprints()}}.}

\item{...}{Additional arguments to \code{\link[=build_prompt_preprint]{build_prompt_preprint()}}.}
}
\value{
A data frame of preprints with a prompt added.
}
\description{
Add prompt to a data frame of preprints
}
\examples{
preprints <- get_preprints(subject=c("bioinformatics", "genomics"))
preprints <- add_prompt(preprints)
preprints

}
\seealso{
\code{\link[=build_prompt_preprint]{build_prompt_preprint()}}
}
