% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{add_prompt_subject}
\alias{add_prompt_subject}
\title{Add prompts for an entire subject}
\usage{
add_prompt_subject(preprints, ...)
}
\arguments{
\item{preprints}{Output from \code{\link[=get_preprints]{get_preprints()}} followed by \code{\link[=add_prompt]{add_prompt()}} followed by \code{\link[=add_summary]{add_summary()}}.}

\item{...}{Additional arguments to \code{\link[=build_prompt_subject]{build_prompt_subject()}}.}
}
\value{
A tibble with a subject and prompt column.
}
\description{
Add prompts for an entire subject
}
\examples{
subjects <-
  example_preprints |>
  dplyr::group_by(subject) |>
  add_prompt_subject()
subjects
}
