% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap-package.R
\docType{package}
\name{biorecap-package}
\alias{biorecap}
\alias{biorecap-package}
\title{biorecap: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
