% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{biorecap_report}
\alias{biorecap_report}
\title{Create a report from bioRxiv/medRxiv preprints}
\usage{
biorecap_report(
  output_dir = ".",
  subject = NULL,
  nsentences = 2L,
  model = "llama3.2",
  host = NULL,
  use_example_preprints = FALSE,
  ...
)
}
\arguments{
\item{output_dir}{Directory to save the report.}

\item{subject}{Character vector of subjects to include in the report.}

\item{nsentences}{Number of sentences to summarize each paper in.}

\item{model}{The model to use for generating summaries. See \code{\link[ollamar:list_models]{ollamar::list_models()}}.}

\item{host}{The base URL to use. Default is \code{NULL}, which uses Ollama's default base URL.}

\item{use_example_preprints}{Use the example preprints data included with the package instead of fetching new data from bioRxiv/medRxiv. For diagnostic/testing purposes only.}

\item{...}{Other arguments passed to \code{\link[rmarkdown:render]{rmarkdown::render()}}.}
}
\value{
Nothing; called for its side effects to produce a report.
}
\description{
Create a report from bioRxiv/medRxiv preprints
}
\examples{
\dontrun{
output_dir <- tempdir()
biorecap_report(use_example_preprints=TRUE, output_dir=output_dir)
biorecap_report(subject=c("bioinformatics", "genomics", "synthetic_biology"),
                output_dir=output_dir)
}
}
