% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{build_prompt_preprint}
\alias{build_prompt_preprint}
\title{Construct a prompt to summarize a paper}
\usage{
build_prompt_preprint(
  title,
  abstract,
  nsentences = 2L,
  instructions = c("I am giving you a paper's title and abstract.",
    "Summarize the paper in as many sentences as I instruct.",
    "Do not include any preamble text to the summary",
    "just give me the summary with no preface or intro sentence.")
)
}
\arguments{
\item{title}{The title of the paper.}

\item{abstract}{The abstract of the paper.}

\item{nsentences}{The number of sentences to summarize the paper in.}

\item{instructions}{Instructions to the prompt. This can be a character vector that gets collapsed into a single string.}
}
\value{
A string containing the prompt.
}
\description{
Construct a prompt to summarize a paper
}
\examples{
build_prompt_preprint(title="A great paper", abstract="This is the abstract.")

}
