% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_preprints}
\alias{example_preprints}
\title{Example preprints with summaries}
\format{
A tibble returned from \code{\link[=get_preprints]{get_preprints()}} followed by \code{\link[=add_prompt]{add_prompt()}} followed by \code{\link[=add_summary]{add_summary()}}.
}
\usage{
example_preprints
}
\description{
Example preprints with summaries from August 6, 2024.
}
\examples{
example_preprints

}
\keyword{datasets}
