% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{safely_query_rss}
\alias{safely_query_rss}
\title{Safely query bioRxiv/medRxiv RSS feeds}
\usage{
safely_query_rss(subject, server = c("biorxiv", "medrxiv"))
}
\arguments{
\item{subject}{A character vector of valid bioRxiv and/or medRxiv subjects. See \link{subjects}.}

\item{server}{A character vector of either "biorxiv" or "medrxiv".}
}
\value{
A data frame of preprints from bioRxiv and/or medRxiv.
}
\description{
Safely query bioRxiv/medRxiv RSS feeds
}
