% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{subjects}
\alias{subjects}
\title{bioRxiv subjects}
\format{
A character vector
}
\usage{
subjects
}
\description{
Names of subjects with RSS feeds in biorXiv
}
\examples{
subjects

}
\references{
\url{https://www.biorxiv.org/alertsrss}
}
\keyword{datasets}
