% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap.R
\name{tt_preprints}
\alias{tt_preprints}
\title{Create a markdown table from prepreprint summaries}
\usage{
tt_preprints(preprints, cols = c("title", "summary"), width = c(1, 3))
}
\arguments{
\item{preprints}{Output from \code{\link[=get_preprints]{get_preprints()}} followed by \code{\link[=add_prompt]{add_prompt()}} followed by \code{\link[=add_summary]{add_summary()}}.}

\item{cols}{Columns to display in the resulting markdown table.}

\item{width}{Vector of relative widths equal to \code{length(cols)}.}
}
\value{
A tinytable table.
}
\description{
Create a markdown table from prepreprint summaries
}
\examples{
# Use built-in example data
example_preprints
tt_preprints(example_preprints[1:2,])
}
