% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biorecap-package.R
\docType{package}
\name{biorecap-package}
\alias{biorecap}
\alias{biorecap-package}
\title{biorecap: Retrieve and summarize bioRxiv preprints with a local LLM using ollama}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Retrieve and summarize bioRxiv preprints with a local LLM using ollama.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://stephenturner.github.io/biorecap/}
  \item Report bugs at \url{https://github.com/stephenturner/biorecap/issues}
}

}
\author{
\strong{Maintainer}: Stephen Turner \email{vustephen@gmail.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})

}
\keyword{internal}
